AllPawnsMustDie is a simple GUI for a chess engine that supports the UCI protocol (e.g. stockfish).  This checkin provides a very meaty skeleton of code to flush out the application.
The main classes and interfaces are defined, and a lot of the basic
routing is hooked up.  The board view will even draw once you load
an engine.

Noteably missing is a lot of the user functionality and the actual launching
and communicating with the chess engine process.  The communication
is proven in the prototype, but also is the logic around the chess
game itelf (like what's a legal move for the piece I clicked on, and
is the player in check, etc)

The next step will likely be to get the process communication working
and have the CPU play itself as it did in the prototype, to finish
testing the drawing and communication hookup.