Play With Yourself(TM) feature is now online along with hookup of the UCI engine and the GUI.  Once an engine is loaded (from the file menu) the user can select File->Self Play to start a game between the engine and itself.  When the game is finished, it will start a new game.  The think time is set fairly low to exercise the communication between the engine and the UI.
The ChessGame and ChessBoard have been updated to account for various
moves that don't come across the UCI protocol explicitly, such as;
castling (moving the rook too) and en-passant captures.  Promotions
are also now handled when coming from the engine.  No UI exists yet
to let the user pick.

The threading model is lacking a bit, but functional.  The engine
currently uses ThreadPool.QueueUserWorkItem to send tasks to the
thread pool in a serialized way, however, only one can go through
at a time right now without breaking the communication.  One is
all we need, so the limit is fine, but enforced only by the calling
code and a bit fragile.

On the view side, moves are now highlighted with colored rects and
lines indicating the last move.  There is a simple text-based
progress indicator to indicate the engine is thinking.