Fixes #3 Enhance Move History
A simple list of strings was sufficient for talking with the engine
over UCI.  However, to make it easier to move backwards, the move
history now contains extra information like, was it a promotion, or
if it was a capture, what piece did it capture.  Walking backwards
can now replace that piece.

Also fixed a few bugs found while developing.  The castling rights
were not being updated correctly due to access (I was updating a
copy).  Also there was a bounds violation when calculating King moves
near the edge of the board.