Fixes # Game Over
Mate is detected when then engine can find no best move.  In every
update case, the code asks for the next bext move (even in the player's
turn - which will be helpful for hints but may be causing occasional
timing issues) so when this happens, the engine fires an event and
a simple messagebox shows the winning player (the one mating).