Fixes #18 Add UnitTests
This change adds UnitTests for several core classes in the AllPawnsMustDie
application.  There is always room to add more, but this is a substantial
starting point.

Missing are the higher level classes, such as the ChessGame, and
the various UI dialogs. Also not included yet are the engine wrapper classes.

Classes covered:
BoardSquare
MoveInformation
ChessPiece
ChessBoard