Fixes #16 Finish up FEN class wrapper
This commit finishes the FEN wrapper class.  It can be initialized
with either a valid FEN string, or with a ChessBoard object.  The
parser will calculate a FEN if given a board, and extract info
needed to construct a board if given a string.  Either way, the
information in the struct is the same in the end, or should be.

UnitTests for the FenParser class are added at the same time and cover
basic loading scenarios, as well as checking an updated fen after
a board move.