Fixes #19 Implement "Show FEN" menu item
As long as a board object exists (you should see one on the screen)
you should be able to get a FEN for that position.  If in self play,
play continues, so you get a snapshot unless you wait until the end.

The dialog itself is very simple, just a readonly TextBox, but this
supports copying to the clipboard, unlike the default MessageBox.

To support waiting until the end, a small 5 second delay has been
added at the end of a self play game.