Fixes #21 Fix thread queueing bug
Long overdue.  As previously noted in the code, the threading model
was very fragile.  The requests were properly serialized to the
engine, but the responses were not properly deserialized, which
could cause the engine to get out of state with the GUI, which
would cause the occassional InvalidOperationStateExection().  This would
often come after a new game on a self play loop.

In the new model, the UCIChessEngine launches a worker thread when
the engine is loaded.  This worker is responsible for the process
and sending/receiving data to it.  IChessEngine.SendCommandAsync now
adds a new command to a shared queue (under lock) and signals the
worker thread.

The worker thread will see new work and send the command.  So far we're
basically at parity pre-change, just implementing portions of ThreadPool
functionality.  What's new, is it will wait - on this thread blocking new
commands from going to the engine - until the response is receveived and
fully processed.

Testing locally, I reduced the game replay timeout from 5 seconds to 10ms
and ran self play in a loop for over 2 hours with no excpetions.  It would
only take at most 10 minutes or so to hit one prior.