Fixes #23 Basic localization
This change adds the core code and setup for localization.  I've added
string resources for German (de-DE) which I'm sure are hilariously
wrong in places.  This was tested locally by setting the UI culture
of the app in code, but should work the same if the UI culture of the
system is de-DE.

Adding additional languages should be a matter of adding new resource
files in the future, and remembering to keep all of the UI strings in
the tables and not the code (minus the language independent ones like
the UCI protocol commands, etc).