Fixes #28 Keep FEN up to date on each move
The FenParser class has undergone a near total rewrite.  For several
reasons (such limitations of engine input buffers) the GUI now updates
its position with the engine only by use of the current FEN.  Bestmoves
are always asked for from a FEN.  The parser is now a stateless set
of static helpers that can extract info from and manipulate a FEN.

The core function is FenParser.ApplyMoveToFEN(fen, san) which given
a current FEN and a move, will produce an updated FEN. The other methods
either support this one, or are used for access to data in the FEN (extract).

Fairly extensive UnitTests were added in conjunction and should be
run whenever this is updated as it is now extremely core to the app.

Fixes #25 Get low hanging fruit for supporting other UCI engines

UCI is a horrible protocol that is not implemented consistently
across engines.  There are many possible results for a "bestmove"
response for example; stockfish returns "bestmove (none)", Rybka
returns "bestmove a1a1", and Spike went with "Error: Fatal no best move"
just to be different.  Some engines crash (Rybkax64) when asked for
a move in a mated position, and others give no reply at all in a mated
position.

I've accounted for a few of these, which manages to cover a lot
more engines at least for normal play.  The ones that won't respond
in a mated position will not properly loop in self play, but may
play fine otherwise.

Here is a quick list of engines I have tried (ad-hoc and quickly)
- StockFish8 (32, 64, bmi, and popcnt)
- Rybka (w32)
- Spike 1.4
- Anmon 5.75

These did NOT work for various reasons with the GUI
- Rybka (x64 - crashes in mated positions)
- SOS51_Arena - not root caused, but partly returns nonstandard bestmove none
- Ruffian_105 - not root caused