Fixes #41, Blocking: Process can leak in Self Play
Two issues addressed in this change.  1) while waiting for the engine
response (which can be on the order of seconds/minutes worst case)
we also wait on the exit event now, in case it comes during that time.
2) The Dispose path for the UCIChessEngine class was just wrong.  At
some point the code diverged and I forgot to clean it up.  Thanks to
stockfish being so damned resilient, I missed it until now.  It's why
testing is important.