design refactor work: postmortem
Issue 1: The Visible property has been renamed to Captured and it's functionally
equivalent to !Visible.  This naming change is to make clear the distinction
between what the piece knows and what the view knows, even though they sort of
overlap in this case.

Issue 2: The Highlight property is purely for the view.  This is information the
ChessPiece should not have to burden itself with and should be decoupled.  In this
case decoupling is acheived via deletion.  The view interface already provides
the needed functionality.