Refactor-postmortem: Implement command pattern for UCI engine commands
This commit adds a command pattern to the UCIChessEngine UCI commands.
Previously, the engine (as well as the ChessGame that contains the engine)
would just send commands to the UCI process "directly" via the IChessEngine
interface (not directly to the process object...).  This leaks knowledge
of UCI implementation and commands/responses when there is no need and
it's better in fact to abstract this out more than it already is.

Goodness from this change:
1)  Decouples the IChessEngine implementation from the command logic
2)  Contains UCI implementation (even how it invokes the interface) from
the caller(s)
3)  Allows direct unit testing of the command sending by replacing the
injected interface IChessEngine
4)  More UnitTests added

Example:
(Old way)
    string position = String.Format("position fen {0}", board.CurrentFEN);
    engine.SendCommandAsync(position, "");

(New Way)
    UciPositionCommand command = new UciPositionCommand(board.CurrentFEN);
    command.Execute(engine);