refactor-postmortem work:
This change decouples the Process object from the UCIChessEngine object.
Access to the engine process is now via IChessEngineProcess, which distills
it down to the raw event and input stream.

The ChessEngineProcessLoader further abstracts the getting and/or creation
of this interface.  The loader is injected to both the ChessGame and
UCIChessEngine, so it can be replaced - which will further improve the
ability to test the ChessGame class.

Before this, I need to decouple the IChessView object from ChessGame
(via injection or other means) so it too can be replaced which is why
tests are not added with this change.