refactor-postmortem: Decouple view from game and more tests
This change has 2 major parts...
PART1: Decouple IChessBoardView from ChessGame object.
The view is now inject in the ChessGame constructor just like the loader.
Now the view is created and owned by the Form, which makes sense, since
the Form is going to be rendering the view anyway.

PART2: More UnitTests and Mocks
With all of the decoupling, more testing is possible, and this change is
mostly about adding supporting pieces as well as tests for those new areas.
Mock implementations for all of the objects decoupled during refactor are
added.  For example, now we can directly test the UCIChessEngine object
by providing a mock loader, process, and stream (directly or indirectly)
so it has no way of knowing it's in a test environment.  We can send
commands to the engine, then intercept what it writes to the process
(because we own the mock) and then respond as if we are the real process.