refactor-postmortem: Decoupling the remaining Form objects
The Form owns the view now, so it also stands that only the view needs
to know about the Form (if anyone, but for the view, yes it needs to know).

ChessGame still had internal knowledge of the form or a control on it.
To fix this, the Invaldate() method was re-added to the IChessBoardView
interface and the 'verbose' control (which really draws the progress
indication right now) is rendered using state set in the view's data
(ChessBoard).

This should open up ChessGame to additional testing options in addition
to just being good decoupling.